--- lib/onetime/app/api/routes
@@ -1,37 +1,0 @@
-
-## API v1
-GET    /v1/status                               Onetime::App::API#status
-
-POST   /v1/share                                Onetime::App::API#share
-POST   /v1/generate                             Onetime::App::API#generate
-
-POST   /v1/create                               Onetime::App::API#create
-
-POST   /v1/secret/:key                          Onetime::App::API#show_secret
-
-GET    /v1/private/recent                       Onetime::App::API#show_metadata_recent
-GET    /v1/metadata/recent                      Onetime::App::API#show_metadata_recent
-
-GET    /v1/private/:key                         Onetime::App::API#show_metadata
-POST   /v1/private/:key                         Onetime::App::API#show_metadata
-GET    /v1/metadata/:key                        Onetime::App::API#show_metadata
-POST   /v1/metadata/:key                        Onetime::App::API#show_metadata
-
-POST   /v1/private/:key/burn                    Onetime::App::API#burn_secret
-
-POST   /v1/account/destroy                      Onetime::App::API::Account#destroy_account
-POST   /v1/account/change-password              Onetime::App::API::Account#change_account_password
-POST   /v1/account/apikey                       Onetime::App::API::Account#generate_apikey
-POST   /v1/account/domains/add                  Onetime::App::API::Domains#add_domain
-POST   /v1/account/domains/:domain/remove       Onetime::App::API::Domains#remove_domain
-GET    /v1/account/domains/:domain              Onetime::App::API::Domains#get_domain
-POST   /v1/account/domains/:domain/verify       Onetime::App::API::Domains#verify_domain
-GET    /v1/account/domains                      Onetime::App::API::Domains#list_domains
-
-## API v2
-GET    /v2/status                               Onetime::App::APIV2#status
-GET    /v2/version                              Onetime::App::APIV2#version
-
-GET    /v2/altcha/challenge                     Onetime::App::APIV2#altcha_challenge
-POST   /v2/altcha/verify                        Onetime::App::APIV2#altcha_verify
-POST   /v2/altcha/verify_spam                   Onetime::App::APIV2#altcha_verify_spam
