+++ lib/onetime/app/api/v2/account.rb
@@ -3,10 +3,10 @@ require_relative 'base'
 require_relative '../../app_settings'
 require_relative '../../../logic/account'
 
-class Onetime::App::API
+class Onetime::App::APIV2
   class Account
     include Onetime::App::AppSettings
-    include Onetime::App::API::Base
+    include Onetime::App::APIV2::Base
 
     @check_utf8 = true
     @check_uri_encoding = true
