+++ lib/onetime/app/api/v2/domains.rb
@@ -4,10 +4,10 @@ require_relative 'base'
 require_relative '../../app_settings'
 require_relative '../../../logic/domains'
 
-class Onetime::App::API
+class Onetime::App::APIV2
   class Domains
     include Onetime::App::AppSettings
-    include Onetime::App::API::Base
+    include Onetime::App::APIV2::Base
 
     @check_utf8 = true
     @check_uri_encoding = true
